Fixed SAT int32 overflow and added blur radius validation (review round)
Fixed a signed-overflow defect found in review across three files
(bgra8888_mean_blur.c from YUV-42, yuv420_box_blur.c and nv21_box_blur.c
from YUV-43): the summed-area-table accumulators were int32_t, and a
single SAT cell at the bottom-right corner sums every sample in the
plane, up to width * height * 255. That exceeds INT32_MAX past roughly
2160p (4096 * 2160 * 255 = 2,256,076,800), so any 4K-or-larger frame
silently overflowed a 32-bit accumulator into undefined behavior and a
wrong blur. Converted the SAT tables, their indices, and the accumulated
per-pixel sums to int64_t in all three files; allocation sizes are
computed through size_t so the calloc call itself cannot overflow
either. The numeric result on existing reference-matrix sizes is
unchanged, since int64_t only widens the safe range where int32_t
previously did not overflow.

Fixed the same review's general finding: boxBlur/meanBlur/gaussianBlur
accepted any radius and passed it straight to native code. A negative
radius inverts the SAT rectangle bounds (x1 > x2, y1 > y2) and produces
an out-of-bounds native read/write; the documented contract in
doc/api-abi-0.3-design.md allows only 0 (no-op) or 1..256. Added
YuvGeometry.validateBlurRadius as a single shared helper and applied it
in both the IO and Web backends' three blur entry points, so validation
does not diverge between them. The IO backend additionally short-circuits
radius == 0 as an explicit Dart-side no-op ahead of any native call or
allocation; the Web backend's existing revision-bump behavior on radius
== 0 was left as is, since changing it is a separate, wider question this
fix does not address.

Added a regression test for the overflow fix: a 3000x2900 all-white
frame, whose single-channel SAT sum (2,218,500,000) already exceeds
INT32_MAX at a fraction of a real 4K buffer's memory, blurred and
checked byte-for-byte against the expected uniform 255. Added a radius
validation test exercising radius -1 and 257 across all three blur
operations, for BGRA in bgra_mean_blur_contract_test.dart and for I420
and NV21 in planar_box_mean_blur_contract_test.dart.

Corrected a reporting error from the prior commit: YUV-43's tracker
entries claimed 10 of 16 planar blur cases passed. An independent
review rerun showed all 16 fail the mae<=4 threshold (4.738 to 9.007),
including the ones previously marked passing — the earlier summary had
compared I420 and NV21 metrics against each other rather than checking
each against the tolerance before writing the summary. No code
regressed between the two measurements; the underlying per-case mae
numbers already recorded in todo-waitlist.md were correct, only the
pass/fail classification built from them was wrong. Widened failure
journal entry F-011 (RGB-domain reference oracle vs YUV-domain blur
operation, not an implementation bug) from the 6 GAUSSIANBLUR-only
cases it previously covered to all 16, and corrected the affected
status lines in todo.md and todo-waitlist.md to RETURNED FOR REWORK
with the actual pass count.

MSVC (VS 2022, /W4 /WX) verified clean on all three overflow-fixed
files. Full suite: 325 passed / 28 failed (was 322/28; +3 new tests).

+++ src/yuv/nv21/nv21_box_blur.c
@@ -4,12 +4,18 @@
  * Inclusive sum over the rectangle [x1, x2] x [y1, y2] of a summed-area table
  * whose rows are `width` entries wide. Mirrors
  * src/yuv/bgra8888/bgra8888_mean_blur.c::yuv_sat_rect.
+ *
+ * Entries are int64_t rather than int32_t: the bottom-right SAT cell sums
+ * every sample in the plane, up to width * height * 255. A 4K luma plane
+ * alone (4096 * 2160 * 255 = 2,256,076,800) already exceeds INT32_MAX, so a
+ * 32-bit accumulator would silently overflow into undefined behavior and a
+ * wrong blur on any 4K-or-larger frame.
  */
-static int32_t yuv_sat_rect(const int32_t *sat, int width, int x1, int y1, int x2, int y2) {
-    int32_t sum = sat[y2 * width + x2];
-    if (y1 > 0) sum -= sat[(y1 - 1) * width + x2];
-    if (x1 > 0) sum -= sat[y2 * width + (x1 - 1)];
-    if (x1 > 0 && y1 > 0) sum += sat[(y1 - 1) * width + (x1 - 1)];
+static int64_t yuv_sat_rect(const int64_t *sat, int width, int x1, int y1, int x2, int y2) {
+    int64_t sum = sat[(int64_t) y2 * width + x2];
+    if (y1 > 0) sum -= sat[(int64_t) (y1 - 1) * width + x2];
+    if (x1 > 0) sum -= sat[(int64_t) y2 * width + (x1 - 1)];
+    if (x1 > 0 && y1 > 0) sum += sat[(int64_t) (y1 - 1) * width + (x1 - 1)];
     return sum;
 }
 
@@ -36,8 +42,8 @@ static void yuv_box_blur_channel(
         int right,
         int bottom
 ) {
-    int32_t *sat = (int32_t *) calloc((size_t) width * height, sizeof(int32_t));
-    uint8_t *temp = (uint8_t *) malloc((size_t) width * height);
+    int64_t *sat = (int64_t *) calloc((size_t) width * (size_t) height, sizeof(int64_t));
+    uint8_t *temp = (uint8_t *) malloc((size_t) width * (size_t) height);
     if (!sat || !temp) {
         free(sat);
         free(temp);
@@ -48,11 +54,11 @@ static void yuv_box_blur_channel(
     }
 
     for (int y = 0; y < height; ++y) {
-        int32_t rowSum = 0;
+        int64_t rowSum = 0;
         for (int x = 0; x < width; ++x) {
             rowSum += temp[y * width + x];
-            const int satIdx = y * width + x;
-            sat[satIdx] = rowSum + (y > 0 ? sat[(y - 1) * width + x] : 0);
+            const int64_t satIdx = (int64_t) y * width + x;
+            sat[satIdx] = rowSum + (y > 0 ? sat[(int64_t) (y - 1) * width + x] : 0);
         }
     }
 
@@ -72,7 +78,7 @@ static void yuv_box_blur_channel(
             const int padLeft = x1 - (x - radius);
             const int padRight = (x + radius) - x2;
 
-            int32_t sum = 0;
+            int64_t sum = 0;
             const struct { int x1, y1, x2, y2, weight; } parts[] = {
                 { x1, y1, x2, y2, 1 },
                 { x1, y1, x2, y1, padTop },
@@ -87,7 +93,7 @@ static void yuv_box_blur_channel(
             for (size_t i = 0; i < sizeof(parts) / sizeof(parts[0]); ++i) {
                 const int weight = parts[i].weight;
                 if (weight == 0) continue;
-                sum += weight * yuv_sat_rect(sat, width, parts[i].x1, parts[i].y1, parts[i].x2, parts[i].y2);
+                sum += (int64_t) weight * yuv_sat_rect(sat, width, parts[i].x1, parts[i].y1, parts[i].x2, parts[i].y2);
             }
 
             temp[y * width + x] = (uint8_t)((sum + half) / area);
